ch(mvc): adopt MVC
