+++ public/client/popup.html
@@ -34,7 +34,10 @@
     </div>
     <div class="footer">
       <span class="clearAll"> Clear All </span>
-      <span class="bugReport"> &#x1F41E; Report a Bug </span>
+      <span> 
+        <span> &#x1F41E;</span>
+        <span class="bugReport"> Report a Bug </span>
+      </span>
     </div>
   </body>
 </html>
