+++ public/src/background.js
@@ -1,1 +1,1 @@
-!function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)n.d(r,o,function(t){return e[t]}.bind(null,o));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=9)}([function(e,t){e.exports=function(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}},function(e,t,n){var r=n(3),o=n(4),c=n(5);e.exports=function(e){return r(e)||o(e)||c()}},function(e,t){function n(e){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function r(t){return"function"==typeof Symbol&&"symbol"===n(Symbol.iterator)?e.exports=r=function(e){return n(e)}:e.exports=r=function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":n(e)},r(t)}e.exports=r},function(e,t){e.exports=function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}},function(e,t){e.exports=function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}},function(e,t){e.exports=function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}},function(e,t,n){"use strict";n.d(t,"a",(function(){return r})),n.d(t,"b",(function(){return o}));var r=function(e,t){return Object.keys(e).every((function(n){return e[n]!==t}))},o={date:(new Date).getUTCDay(),delay:7,theme:"light",autoSave:"no"}},function(e,t){e.exports=function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}},function(e,t){function n(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}e.exports=function(e,t,r){return t&&n(e.prototype,t),r&&n(e,r),e}},function(e,t,n){"use strict";n.r(t);var r=n(0),o=n.n(r),c=n(1),i=n.n(c),u=n(2),s=n.n(u),a=n(7),l=n.n(a),f=n(8),p=n.n(f),y=n(6);function b(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(e);t&&(r=r.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,r)}return n}function m(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?b(n,!0).forEach((function(t){o()(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):b(n).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}chrome.runtime.onInstalled.addListener((function(){chrome.storage.sync.get("clips",(function(e){Object.keys(e).length>0&&e.clips?chrome.browserAction.setBadgeText({text:"".concat(Object.keys(e.clips).length)}):chrome.browserAction.setBadgeText({text:""})})),chrome.storage.sync.get("settings",(function(e){0===Object.keys(e).length&&chrome.storage.sync.set({settings:y.b})})),chrome.browserAction.setBadgeBackgroundColor({color:"#4688F1"})})),chrome.contextMenus.create({id:"selection",title:"Copy to MultiClip",contexts:["selection","link","image"]}),chrome.contextMenus.create({id:"page",title:"About MultiClip",contexts:["page","browser_action"]}),chrome.contextMenus.create({id:"clearClippedItems",title:"Clear Clipped Items",contexts:["page","browser_action"]});var g=function(){function e(){l()(this,e)}return p()(e,null,[{key:"selection",value:function(e){var t=e.selectionText;chrome.storage.sync.get("clips",(function(e){var n,r,c=e.clips,u="object"===s()(c),a=t.trim(),l=""!==t.trim();if(u&&l&&(n=c,r=t,Object.keys(n).every((function(e){return n[e]!==r})))){var f=Math.max.apply(Math,i()(Object.keys(c))),p=m({},e.clips,o()({},f+1,a));chrome.storage.sync.set({clips:p})}else!u&&l&&chrome.storage.sync.set({clips:{1:t}})}))}},{key:"page",value:function(){chrome.tabs.create({url:"https://github.com/Oluwasegun-AA/MultiClip"})}},{key:"lightTheme",value:function(){return chrome.storage.sync.get("settings",(function(e){var t=m({},e.settings,{theme:"light"});chrome.storage.sync.set({settings:t})}))}},{key:"darkTheme",value:function(){return chrome.storage.sync.get("settings",(function(e){var t=m({},e.settings,{theme:"dark"});chrome.storage.sync.set({settings:t})}))}},{key:"deleteItems",value:function(){}},{key:"clearClippedItems",value:function(){return chrome.storage.sync.remove("clips")}}]),e}();chrome.contextMenus.onClicked.addListener((function(e,t){var n=e.menuItemId;g[n](e,t)}))}]);
+!function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)n.d(r,o,function(t){return e[t]}.bind(null,o));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=12)}([function(e,t,n){"use strict";n.d(t,"a",(function(){return a})),n.d(t,"b",(function(){return f})),n.d(t,"f",(function(){return y})),n.d(t,"g",(function(){return p})),n.d(t,"c",(function(){return b})),n.d(t,"d",(function(){return d})),n.d(t,"e",(function(){return g}));var r=n(1),o=n.n(r),c=n(2),i=n.n(c),u=n(3),l=n.n(u);function s(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(e);t&&(r=r.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,r)}return n}var a=function(e,t){return Object.keys(e).every((function(n){return e[n]!==t}))},f={date:(new Date).getUTCDay(),delay:7,theme:"light",autoSave:"no"},y=function(e){chrome.storage.sync.get("clips",(function(t){var n=t.clips,r="object"===l()(n),c=e.trim(),u=""!==c;if(r&&u&&a(t.clips,e)){var f=Math.max.apply(Math,i()(Object.keys(n))),y=function(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?s(n,!0).forEach((function(t){o()(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):s(n).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}({},t.clips,o()({},f+1,c));chrome.storage.sync.set({clips:y})}else if(!r&&u){var p={1:e};chrome.storage.sync.set({clips:p})}}))},p=function(e){chrome.browserAction.setBadgeText({text:"".concat(e)})},b=function(e){return chrome.i18n.getMessage(e)},d=function(e,t,n,r){"dark"===e?(t.style.backgroundColor="#2f2d2d",t.style.color="#d6d3d3",n.forEach((function(e){e.style.backgroundColor="#2f2d2d"})),r.forEach((function(e){e.style.backgroundColor="#2f2d2d",e.style.border="none",e.style.color="#d6d3d3"}))):(t.style.backgroundColor="white",t.style.color="black",n.forEach((function(e){e.style.backgroundColor="white"})),r.forEach((function(e){e.style.backgroundColor="white",e.style.color="black"})))},g=function(e){chrome.tabs.create({url:e})}},function(e,t){e.exports=function(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}},function(e,t,n){var r=n(4),o=n(5),c=n(6);e.exports=function(e){return r(e)||o(e)||c()}},function(e,t){function n(e){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function r(t){return"function"==typeof Symbol&&"symbol"===n(Symbol.iterator)?e.exports=r=function(e){return n(e)}:e.exports=r=function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":n(e)},r(t)}e.exports=r},function(e,t){e.exports=function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}},function(e,t){e.exports=function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}},function(e,t){e.exports=function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}},function(e,t){e.exports=function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}},function(e,t){function n(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}e.exports=function(e,t,r){return t&&n(e.prototype,t),r&&n(e,r),e}},,,,function(e,t,n){"use strict";n.r(t);var r=n(0),o=n(7),c=n.n(o),i=n(8),u=n.n(i),l=n(1),s=n.n(l);function a(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(e);t&&(r=r.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,r)}return n}var f=function(e){chrome.storage.sync.get("settings",(function(t){var n=function(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?a(n,!0).forEach((function(t){s()(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):a(n).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}({},t.settings,{theme:e});chrome.storage.sync.set({settings:n})}))},y=function(){function e(){c()(this,e)}return u()(e,null,[{key:"selection",value:function(e){var t=e.selectionText;Object(r.f)(t)}},{key:"page",value:function(){Object(r.e)("https://github.com/Oluwasegun-AA/MultiClip")}},{key:"lightTheme",value:function(){return f("light")}},{key:"darkTheme",value:function(){return f("dark")}},{key:"deleteItems",value:function(){}},{key:"clearClippedItems",value:function(){return chrome.storage.sync.remove("clips")}}]),e}();chrome.runtime.onInstalled.addListener((function(){chrome.storage.sync.get("clips",(function(e){Object.keys(e).length>0&&e.clips?Object(r.g)("".concat(Object.keys(e.clips).length)):Object(r.g)("")})),chrome.storage.sync.get("settings",(function(e){0===Object.keys(e).length&&chrome.storage.sync.set({settings:r.b})})),chrome.browserAction.setBadgeBackgroundColor({color:"#4688F1"})})),chrome.contextMenus.create({id:"selection",title:"Copy to MultiClip",contexts:["selection","link","image"]}),chrome.contextMenus.create({id:"page",title:"About MultiClip",contexts:["page","browser_action"]}),chrome.contextMenus.create({id:"clearClippedItems",title:"Clear Clipped Items",contexts:["page","browser_action"]}),chrome.contextMenus.onClicked.addListener((function(e,t){var n=e.menuItemId;y[n](e,t)}))}]);
