+++ public/src/contentScript.js
@@ -1,1 +1,1 @@
-!function(t){var e={};function r(n){if(e[n])return e[n].exports;var o=e[n]={i:n,l:!1,exports:{}};return t[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}r.m=t,r.c=e,r.d=function(t,e,n){r.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},r.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},r.t=function(t,e){if(1&e&&(t=r(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)r.d(n,o,function(e){return t[e]}.bind(null,o));return n},r.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return r.d(e,"a",e),e},r.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},r.p="",r(r.s=10)}([function(t,e){t.exports=function(t,e,r){return e in t?Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[e]=r,t}},function(t,e,r){var n=r(3),o=r(4),c=r(5);t.exports=function(t){return n(t)||o(t)||c()}},function(t,e){function r(t){return(r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function n(e){return"function"==typeof Symbol&&"symbol"===r(Symbol.iterator)?t.exports=n=function(t){return r(t)}:t.exports=n=function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":r(t)},n(e)}t.exports=n},function(t,e){t.exports=function(t){if(Array.isArray(t)){for(var e=0,r=new Array(t.length);e<t.length;e++)r[e]=t[e];return r}}},function(t,e){t.exports=function(t){if(Symbol.iterator in Object(t)||"[object Arguments]"===Object.prototype.toString.call(t))return Array.from(t)}},function(t,e){t.exports=function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}},function(t,e,r){"use strict";r.d(e,"a",(function(){return n})),r.d(e,"b",(function(){return o}));var n=function(t,e){return Object.keys(t).every((function(r){return t[r]!==e}))},o={date:(new Date).getUTCDay(),delay:7,theme:"light",autoSave:"no"}},,,,function(t,e,r){"use strict";r.r(e);var n=r(0),o=r.n(n),c=r(1),i=r.n(c),u=r(2),f=r.n(u),a=r(6);function l(t,e){var r=Object.keys(t);if(Object.getOwnPropertySymbols){var n=Object.getOwnPropertySymbols(t);e&&(n=n.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),r.push.apply(r,n)}return r}var s=function(){chrome.storage.sync.get("settings",(function(t){if("yes"===t.settings.autoSave)return navigator.clipboard.readText().then((function(t){chrome.storage.sync.get("clips",(function(e){var r=e.clips,n="object"===f()(r),c=t.trim(),u=""!==c;if(n&&u&&Object(a.a)(e.clips,t)){var s=Math.max.apply(Math,i()(Object.keys(r))),p=function(t){for(var e=1;e<arguments.length;e++){var r=null!=arguments[e]?arguments[e]:{};e%2?l(r,!0).forEach((function(e){o()(t,e,r[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(r)):l(r).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(r,e))}))}return t}({},e.clips,o()({},s+1,c));chrome.storage.sync.set({clips:p})}else if(!n&&u){var y={1:t};chrome.storage.sync.set({clips:y})}}))})).catch((function(t){console.warn("You need to authorize the page to access your clipboard in order to copy to multiClip",t)}))}))};document.addEventListener("copy",(function(){s()}))}]);
+!function(t){var e={};function r(n){if(e[n])return e[n].exports;var o=e[n]={i:n,l:!1,exports:{}};return t[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}r.m=t,r.c=e,r.d=function(t,e,n){r.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},r.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},r.t=function(t,e){if(1&e&&(t=r(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)r.d(n,o,function(e){return t[e]}.bind(null,o));return n},r.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return r.d(e,"a",e),e},r.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},r.p="",r(r.s=9)}([function(t,e,r){"use strict";r.d(e,"a",(function(){return l})),r.d(e,"b",(function(){return s})),r.d(e,"f",(function(){return y})),r.d(e,"g",(function(){return p})),r.d(e,"c",(function(){return b})),r.d(e,"d",(function(){return d})),r.d(e,"e",(function(){return g}));var n=r(1),o=r.n(n),c=r(2),u=r.n(c),i=r(3),f=r.n(i);function a(t,e){var r=Object.keys(t);if(Object.getOwnPropertySymbols){var n=Object.getOwnPropertySymbols(t);e&&(n=n.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),r.push.apply(r,n)}return r}var l=function(t,e){return Object.keys(t).every((function(r){return t[r]!==e}))},s={date:(new Date).getUTCDay(),delay:7,theme:"light",autoSave:"no"},y=function(t){chrome.storage.sync.get("clips",(function(e){var r=e.clips,n="object"===f()(r),c=t.trim(),i=""!==c;if(n&&i&&l(e.clips,t)){var s=Math.max.apply(Math,u()(Object.keys(r))),y=function(t){for(var e=1;e<arguments.length;e++){var r=null!=arguments[e]?arguments[e]:{};e%2?a(r,!0).forEach((function(e){o()(t,e,r[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(r)):a(r).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(r,e))}))}return t}({},e.clips,o()({},s+1,c));chrome.storage.sync.set({clips:y})}else if(!n&&i){var p={1:t};chrome.storage.sync.set({clips:p})}}))},p=function(t){chrome.browserAction.setBadgeText({text:"".concat(t)})},b=function(t){return chrome.i18n.getMessage(t)},d=function(t,e,r,n){"dark"===t?(e.style.backgroundColor="#2f2d2d",e.style.color="#d6d3d3",r.forEach((function(t){t.style.backgroundColor="#2f2d2d"})),n.forEach((function(t){t.style.backgroundColor="#2f2d2d",t.style.border="none",t.style.color="#d6d3d3"}))):(e.style.backgroundColor="white",e.style.color="black",r.forEach((function(t){t.style.backgroundColor="white"})),n.forEach((function(t){t.style.backgroundColor="white",t.style.color="black"})))},g=function(t){chrome.tabs.create({url:t})}},function(t,e){t.exports=function(t,e,r){return e in t?Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[e]=r,t}},function(t,e,r){var n=r(4),o=r(5),c=r(6);t.exports=function(t){return n(t)||o(t)||c()}},function(t,e){function r(t){return(r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function n(e){return"function"==typeof Symbol&&"symbol"===r(Symbol.iterator)?t.exports=n=function(t){return r(t)}:t.exports=n=function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":r(t)},n(e)}t.exports=n},function(t,e){t.exports=function(t){if(Array.isArray(t)){for(var e=0,r=new Array(t.length);e<t.length;e++)r[e]=t[e];return r}}},function(t,e){t.exports=function(t){if(Symbol.iterator in Object(t)||"[object Arguments]"===Object.prototype.toString.call(t))return Array.from(t)}},function(t,e){t.exports=function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}},,,function(t,e,r){"use strict";r.r(e);var n=r(0),o=function(){chrome.storage.sync.get("settings",(function(t){if("yes"===t.settings.autoSave)return e=n.f,void navigator.clipboard.readText().then((function(t){e(t)})).catch((function(t){console.warn("You need to authorize the page to access your clipboard in order to copy to multiClip",t)}));var e}))};document.addEventListener("copy",(function(){o()}))}]);
