+++ public/src/popup.js
@@ -1,1 +1,1 @@
-!function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)n.d(r,o,function(t){return e[t]}.bind(null,o));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=11)}([function(e,t){e.exports=function(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}},function(e,t,n){var r=n(3),o=n(4),c=n(5);e.exports=function(e){return r(e)||o(e)||c()}},function(e,t){function n(e){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function r(t){return"function"==typeof Symbol&&"symbol"===n(Symbol.iterator)?e.exports=r=function(e){return n(e)}:e.exports=r=function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":n(e)},r(t)}e.exports=r},function(e,t){e.exports=function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}},function(e,t){e.exports=function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}},function(e,t){e.exports=function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}},,,,,,function(e,t,n){"use strict";n.r(t);var r=n(1),o=n.n(r),c=n(2),i=n.n(c),s=n(0),l=n.n(s);function a(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(e);t&&(r=r.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,r)}return n}function u(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?a(n,!0).forEach((function(t){l()(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):a(n).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}var p=document.querySelector.bind(document),y=document.querySelectorAll.bind(document),f=p(".emptyPrompt"),d=p(".logo"),b=p(".clearAll"),m=p(".clipText"),g=p(".clips"),v=p(".clips"),h=p(".hint"),O=p(".copyPrompt"),j=p(".hintText"),k=p(".text"),x=p(".bugReport"),S=p("body"),P=y(".bugReport, .clearAll, .emptyPrompt, .copyPrompt, .logo, .settings"),E=y(".footer, .header"),w=y(".settings, .clearAll, .bugReport, .emptyPrompt, .copyPrompt"),T=function(e){chrome.browserAction.setBadgeText({text:"".concat(e)})},M=function(){h.style.display="none",f.style.display="block",O.style.display="block",m.style.display="none",chrome.storage.sync.remove("clips"),g.textContent="",b.style.visibility="hidden",T("")};w.forEach((function(e){var t,n=e.className;e.innerHTML=(t=n,chrome.i18n.getMessage(t))}));var A=function(e){var t="";Object.keys(e).forEach((function(n){t+='\n            <div id="no'.concat(n,'" name="').concat(n,'" class="clip">\n            <span class="copy">\n            &#x2398;\n            </span>\n            <span id="no').concat(n,'" class="content">\n            ').concat(e["".concat(n)],'\n            </span>\n            <span class="deleteClip"> &#x274C; </span>\n            </div>')})),g.innerHTML=t};chrome.storage.sync.get("clips",(function(e){if(e.clips&&0!==Object.keys(e.clips).length){var t=Object.keys(e.clips).length;A(e.clips),T(t)}else f.style.display="block",b.style.visibility="hidden",h.style.display="none",T("")}));var L=function(e,t,n){n>=t&&(chrome.storage.sync.set({settings:u({},e,{date:(new Date).getUTCDay()})}),chrome.storage.sync.remove("clips"),M())};chrome.storage.sync.get("settings",(function(e){var t=e.settings,n=t.delay,r=t.date;!function(e){"dark"===e&&(S.style.backgroundColor="#2f2d2d",E.forEach((function(e){e.style.backgroundColor="#2f2d2d"})),P.forEach((function(e){e.style.backgroundColor="#2f2d2d",e.style.border="none",e.style.color="#d6d3d3"})),S.style.color="#d6d3d3")}(t.theme);var o=(new Date).getUTCDay();if(0!==n&&o!==r)if(o<r){var c=r-Math.abs(o-r);L(e.settings,n,c)}else{var i=o-r;L(e.settings,n,i)}})),p(".prompt").addEventListener("click",(function(){var e=p(".copyPrompt"),t=p(".clipText");e.style.display="none",t.style.display="block",t.focus()}));p(".clipText").addEventListener("keypress",(function(e){var t=e.keyCode,n=e.shiftKey;13===t&&!0!==n&&(e.preventDefault(),chrome.storage.sync.get("clips",(function(e){var t,n,r=e.clips,c="object"===i()(r),s=m.value.trim(),a=""!==s;if(c&&a&&(t=e.clips,n=s,Object.keys(t).every((function(e){return t[e]!==n})))){var p=Math.max.apply(Math,o()(Object.keys(r))),y=u({},e.clips,l()({},p+1,s));chrome.storage.sync.set({clips:y},(function(){A(y)})),T(Object.keys(y).length)}else if(!c&&a){var d={1:s};chrome.storage.sync.set({clips:d},(function(){A(d)}))}m.value="",b.style.visibility="visible",m.focus(),f.style.display="none",h.style.display="block"})))})),v.addEventListener("mouseover",(function(e){"content"===e.target.className&&(j.style.display="none",k.style.display="block",k.innerHTML=e.target.innerText.trim())})),v.addEventListener("mouseout",(function(){j.style.display="block",k.style.display="none"})),b.addEventListener("click",(function(){M()})),v.addEventListener("click",(function(e){if("deleteClip"===e.target.className){var t=e.target.previousElementSibling.id,n={};chrome.storage.sync.get("clips",(function(e){var r=t.substr(2),o=Object.keys(e.clips).filter((function(e){return e!==r}));o.forEach((function(t){n=u({},n,l()({},"".concat(t),e.clips["".concat(t)]))})),chrome.storage.sync.set({clips:n});var c=o.length;0===c?M():(A(n),T(c))}))}}));var C=function(e){navigator.clipboard.writeText(e)};v.addEventListener("click",(function(e){var t=e.target.className;"content"===t?C(e.target.innerHTML.trim()):"copy"===t&&C(e.target.nextElementSibling.innerHTML.trim())})),x.addEventListener("click",(function(){chrome.tabs.create({url:"https://github.com/Oluwasegun-AA/MultiClip/issues"})})),d.addEventListener("click",(function(){chrome.tabs.create({url:"https://github.com/Oluwasegun-AA/MultiClip"})}))}]);
+!function(t){var e={};function n(r){if(e[r])return e[r].exports;var o=e[r]={i:r,l:!1,exports:{}};return t[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)n.d(r,o,function(e){return t[e]}.bind(null,o));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=10)}([function(t,e,n){"use strict";n.d(e,"a",(function(){return u})),n.d(e,"b",(function(){return p})),n.d(e,"f",(function(){return y})),n.d(e,"g",(function(){return f})),n.d(e,"c",(function(){return b})),n.d(e,"d",(function(){return d})),n.d(e,"e",(function(){return g}));var r=n(1),o=n.n(r),c=n(2),i=n.n(c),s=n(3),l=n.n(s);function a(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}var u=function(t,e){return Object.keys(t).every((function(n){return t[n]!==e}))},p={date:(new Date).getUTCDay(),delay:7,theme:"light",autoSave:"no"},y=function(t){chrome.storage.sync.get("clips",(function(e){var n=e.clips,r="object"===l()(n),c=t.trim(),s=""!==c;if(r&&s&&u(e.clips,t)){var p=Math.max.apply(Math,i()(Object.keys(n))),y=function(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?a(n,!0).forEach((function(e){o()(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):a(n).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}({},e.clips,o()({},p+1,c));chrome.storage.sync.set({clips:y})}else if(!r&&s){var f={1:t};chrome.storage.sync.set({clips:f})}}))},f=function(t){chrome.browserAction.setBadgeText({text:"".concat(t)})},b=function(t){return chrome.i18n.getMessage(t)},d=function(t,e,n,r){"dark"===t?(e.style.backgroundColor="#2f2d2d",e.style.color="#d6d3d3",n.forEach((function(t){t.style.backgroundColor="#2f2d2d"})),r.forEach((function(t){t.style.backgroundColor="#2f2d2d",t.style.border="none",t.style.color="#d6d3d3"}))):(e.style.backgroundColor="white",e.style.color="black",n.forEach((function(t){t.style.backgroundColor="white"})),r.forEach((function(t){t.style.backgroundColor="white",t.style.color="black"})))},g=function(t){chrome.tabs.create({url:t})}},function(t,e){t.exports=function(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}},function(t,e,n){var r=n(4),o=n(5),c=n(6);t.exports=function(t){return r(t)||o(t)||c()}},function(t,e){function n(t){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function r(e){return"function"==typeof Symbol&&"symbol"===n(Symbol.iterator)?t.exports=r=function(t){return n(t)}:t.exports=r=function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":n(t)},r(e)}t.exports=r},function(t,e){t.exports=function(t){if(Array.isArray(t)){for(var e=0,n=new Array(t.length);e<t.length;e++)n[e]=t[e];return n}}},function(t,e){t.exports=function(t){if(Symbol.iterator in Object(t)||"[object Arguments]"===Object.prototype.toString.call(t))return Array.from(t)}},function(t,e){t.exports=function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}},,,,function(t,e,n){"use strict";n.r(e);var r=n(2),o=n.n(r),c=n(3),i=n.n(c),s=n(1),l=n.n(s),a=n(0);function u(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function p(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?u(n,!0).forEach((function(e){l()(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):u(n).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}var y=document.querySelector.bind(document),f=document.querySelectorAll.bind(document),b=y(".emptyPrompt"),d=y(".logo"),g=y(".clearAll"),m=y(".clipText"),v=y(".clips"),h=y(".clips"),O=y(".hint"),j=y(".copyPrompt"),k=y(".hintText"),P=y(".text"),x=y(".bugReport"),w=y("body"),S=f(".bugReport, .clearAll, .emptyPrompt, .copyPrompt, .logo, .settings"),E=f(".footer, .header"),T=f(".settings, .clearAll, .bugReport, .emptyPrompt, .copyPrompt"),M=function(){O.style.display="none",b.style.display="block",j.style.display="block",m.style.display="none",chrome.storage.sync.remove("clips"),v.textContent="",g.style.visibility="hidden",Object(a.g)("")};T.forEach((function(t){var e=t.className;t.innerHTML=Object(a.c)(e)}));var C=function(t){var e="";Object.keys(t).forEach((function(n){e+='\n            <div id="no'.concat(n,'" name="').concat(n,'" class="clip">\n            <span class="copy">\n            &#x2398;\n            </span>\n            <span id="no').concat(n,'" class="content">\n            ').concat(t["".concat(n)],'\n            </span>\n            <span class="deleteClip"> &#x274C; </span>\n            </div>')})),v.innerHTML=e};chrome.storage.sync.get("clips",(function(t){if(t.clips&&0!==Object.keys(t.clips).length){var e=Object.keys(t.clips).length;C(t.clips),Object(a.g)(e)}else b.style.display="block",g.style.visibility="hidden",O.style.display="none",Object(a.g)("")}));var D=function(t,e,n){n>=e&&(chrome.storage.sync.set({settings:p({},t,{date:(new Date).getUTCDay()})}),chrome.storage.sync.remove("clips"),M())};chrome.storage.sync.get("settings",(function(t){var e=t.settings,n=e.delay,r=e.date,o=e.theme;Object(a.d)(o,w,E,S);var c=(new Date).getUTCDay();if(0!==n&&c!==r)if(c<r){var i=r-Math.abs(c-r);D(t.settings,n,i)}else{var s=c-r;D(t.settings,n,s)}})),y(".prompt").addEventListener("click",(function(){var t=y(".copyPrompt"),e=y(".clipText");t.style.display="none",e.style.display="block",e.focus()})),y(".clipText").addEventListener("keypress",(function(t){var e=t.keyCode,n=t.shiftKey;13===e&&!0!==n&&(t.preventDefault(),chrome.storage.sync.get("clips",(function(t){var e=t.clips,n="object"===i()(e),r=m.value.trim(),c=""!==r;if(n&&c&&Object(a.a)(t.clips,r)){var s=Math.max.apply(Math,o()(Object.keys(e))),u=p({},t.clips,l()({},s+1,r));chrome.storage.sync.set({clips:u},(function(){C(u)})),Object(a.g)(Object.keys(u).length)}else if(!n&&c){var y={1:r};chrome.storage.sync.set({clips:y},(function(){C(y)}))}m.value="",g.style.visibility="visible",m.focus(),b.style.display="none",O.style.display="block"})))})),h.addEventListener("mouseover",(function(t){"content"===t.target.className&&(k.style.display="none",P.style.display="block",P.innerHTML=t.target.innerText.trim())})),h.addEventListener("mouseout",(function(){k.style.display="block",P.style.display="none"})),g.addEventListener("click",(function(){M()})),h.addEventListener("click",(function(t){if("deleteClip"===t.target.className){var e=t.target.previousElementSibling.id,n={};chrome.storage.sync.get("clips",(function(t){var r=e.substr(2),o=Object.keys(t.clips).filter((function(t){return t!==r}));o.forEach((function(e){n=p({},n,l()({},"".concat(e),t.clips["".concat(e)]))})),chrome.storage.sync.set({clips:n});var c=o.length;0===c?M():(C(n),Object(a.g)(c))}))}}));var A=function(t){navigator.clipboard.writeText(t)};h.addEventListener("click",(function(t){var e=t.target.className;"content"===e?A(t.target.innerHTML.trim()):"copy"===e&&A(t.target.nextElementSibling.innerHTML.trim())})),x.addEventListener("click",(function(){Object(a.e)("https://github.com/Oluwasegun-AA/MultiClip/issues")})),d.addEventListener("click",(function(){Object(a.e)("https://github.com/Oluwasegun-AA/MultiClip")}))}]);
