+++ public/src/settings.js
@@ -1,1 +1,1 @@
-!function(e){var t={};function n(o){if(t[o])return t[o].exports;var r=t[o]={i:o,l:!1,exports:{}};return e[o].call(r.exports,r,r.exports,n),r.l=!0,r.exports}n.m=e,n.c=t,n.d=function(e,t,o){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:o})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var o=Object.create(null);if(n.r(o),Object.defineProperty(o,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var r in e)n.d(o,r,function(t){return e[t]}.bind(null,r));return o},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=12)}({12:function(e,t){var n=document.querySelector.bind(document),o=document.querySelectorAll.bind(document),r=n("#delay"),l=n("#theme"),a=n("#autoSave"),c=n(".saveButton"),s=n(".saveDiv"),u=n(".warningDiv"),i=n(".restoreDefault"),d=n(".bugReport"),f=n("body"),y=o(".bugReport, .restoreDefault, .backBtn, .options, .settings, .tag"),g=o(".footer, .header"),b=o(".settings, .note, .noteText, .bugReport, .saveButton, .restoreDefault, .no1, .no2, .no3, .no4, .no5, .no6, .no7, .no8, .no9, .no0, .dark, .light, .yes, .no");b.forEach((function(e){var t,n=e.className;e.innerHTML=(t=n,chrome.i18n.getMessage(t))}));var p=function(e){r.value=e.delay,l.value=e.theme,a.value=e.autoSave},v=function(e){"dark"===e?(f.style.backgroundColor="#2f2d2d",f.style.color="#d6d3d3",g.forEach((function(e){e.style.backgroundColor="#2f2d2d"})),y.forEach((function(e){e.style.backgroundColor="#2f2d2d",e.style.border="none",e.style.color="#d6d3d3"}))):(f.style.backgroundColor="white",f.style.color="black",g.forEach((function(e){e.style.backgroundColor="white"})),y.forEach((function(e){e.style.backgroundColor="white",e.style.color="black"})))};chrome.storage.sync.get("settings",(function(e){p(e.settings),v(e.settings.theme)}));c.addEventListener("click",(function(){var e,t,n,o,c,i,d;chrome.storage.sync.set({settings:(e=r.options,t=r.selectedIndex,n=l.options,o=l.selectedIndex,c=a.options,i=a.selectedIndex,d=parseInt(e[t].value,10),{date:d>0?(new Date).getUTCDay():"",delay:d,theme:n[o].value,autoSave:c[i].value})},(function(){chrome.storage.sync.get("settings",(function(e){v(e.settings.theme)}))})),s.style.display="none",u.style.display="none"})),o("select").forEach((function(e){e.addEventListener("change",(function(){s.style.display="block"}))})),i.addEventListener("click",(function(){p({delay:7,theme:"light",autoSave:"no"}),s.style.display="block"})),d.addEventListener("click",(function(){chrome.tabs.create({url:"https://github.com/Oluwasegun-AA/MultiClip/issues"})})),a.addEventListener("change",(function(){0===a.selectedIndex?u.style.display="block":u.style.display="none"}))}});
+!function(e){var t={};function n(o){if(t[o])return t[o].exports;var r=t[o]={i:o,l:!1,exports:{}};return e[o].call(r.exports,r,r.exports,n),r.l=!0,r.exports}n.m=e,n.c=t,n.d=function(e,t,o){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:o})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var o=Object.create(null);if(n.r(o),Object.defineProperty(o,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var r in e)n.d(o,r,function(t){return e[t]}.bind(null,r));return o},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=11)}([function(e,t,n){"use strict";n.d(t,"a",(function(){return l})),n.d(t,"b",(function(){return f})),n.d(t,"f",(function(){return y})),n.d(t,"g",(function(){return d})),n.d(t,"c",(function(){return p})),n.d(t,"d",(function(){return b})),n.d(t,"e",(function(){return g}));var o=n(1),r=n.n(o),c=n(2),i=n.n(c),u=n(3),s=n.n(u);function a(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var o=Object.getOwnPropertySymbols(e);t&&(o=o.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,o)}return n}var l=function(e,t){return Object.keys(e).every((function(n){return e[n]!==t}))},f={date:(new Date).getUTCDay(),delay:7,theme:"light",autoSave:"no"},y=function(e){chrome.storage.sync.get("clips",(function(t){var n=t.clips,o="object"===s()(n),c=e.trim(),u=""!==c;if(o&&u&&l(t.clips,e)){var f=Math.max.apply(Math,i()(Object.keys(n))),y=function(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?a(n,!0).forEach((function(t){r()(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):a(n).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}({},t.clips,r()({},f+1,c));chrome.storage.sync.set({clips:y})}else if(!o&&u){var d={1:e};chrome.storage.sync.set({clips:d})}}))},d=function(e){chrome.browserAction.setBadgeText({text:"".concat(e)})},p=function(e){return chrome.i18n.getMessage(e)},b=function(e,t,n,o){"dark"===e?(t.style.backgroundColor="#2f2d2d",t.style.color="#d6d3d3",n.forEach((function(e){e.style.backgroundColor="#2f2d2d"})),o.forEach((function(e){e.style.backgroundColor="#2f2d2d",e.style.border="none",e.style.color="#d6d3d3"}))):(t.style.backgroundColor="white",t.style.color="black",n.forEach((function(e){e.style.backgroundColor="white"})),o.forEach((function(e){e.style.backgroundColor="white",e.style.color="black"})))},g=function(e){chrome.tabs.create({url:e})}},function(e,t){e.exports=function(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}},function(e,t,n){var o=n(4),r=n(5),c=n(6);e.exports=function(e){return o(e)||r(e)||c()}},function(e,t){function n(e){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function o(t){return"function"==typeof Symbol&&"symbol"===n(Symbol.iterator)?e.exports=o=function(e){return n(e)}:e.exports=o=function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":n(e)},o(t)}e.exports=o},function(e,t){e.exports=function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}},function(e,t){e.exports=function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}},function(e,t){e.exports=function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}},,,,,function(e,t,n){"use strict";n.r(t);var o=n(0),r=document.querySelector.bind(document),c=document.querySelectorAll.bind(document),i=r("#delay"),u=r("#theme"),s=r("#autoSave"),a=r(".saveButton"),l=r(".saveDiv"),f=r(".warningDiv"),y=r(".restoreDefault"),d=r(".bugReport"),p=r("body"),b=c(".bugReport, .restoreDefault, .backBtn, .options, .settings, .tag"),g=c(".footer, .header"),m=c(".settings, .note, .noteText, .bugReport, .saveButton, .restoreDefault, .no1, .no2, .no3, .no4, .no5, .no6, .no7, .no8, .no9, .no0, .dark, .light, .yes, .no");m.forEach((function(e){var t,n=e.className;e.innerHTML=(t=n,chrome.i18n.getMessage(t))}));var v=function(e){i.value=e.delay,u.value=e.theme,s.value=e.autoSave};chrome.storage.sync.get("settings",(function(e){var t=e.settings.theme;v(e.settings),Object(o.d)(t,p,g,b)}));a.addEventListener("click",(function(){var e,t,n,r,c,a,y;chrome.storage.sync.set({settings:(e=i.options,t=i.selectedIndex,n=u.options,r=u.selectedIndex,c=s.options,a=s.selectedIndex,y=parseInt(e[t].value,10),{date:y>0?(new Date).getUTCDay():"",delay:y,theme:n[r].value,autoSave:c[a].value})},(function(){chrome.storage.sync.get("settings",(function(e){var t=e.settings.theme;Object(o.d)(t,p,g,b)}))})),l.style.display="none",f.style.display="none"})),c("select").forEach((function(e){e.addEventListener("change",(function(){l.style.display="block"}))})),y.addEventListener("click",(function(){v(o.b),l.style.display="block"})),d.addEventListener("click",(function(){Object(o.e)("https://github.com/Oluwasegun-AA/MultiClip/issues")})),s.addEventListener("change",(function(){0===s.selectedIndex?f.style.display="block":f.style.display="none"}))}]);
