+++ src/app/background.js
@@ -1,14 +1,13 @@
-import { defaultSettings } from '../common/index';
+import { defaultSettings, showBadge } from '../common/index';
+import ContextMenuManager from './utils';
 
 chrome.runtime.onInstalled.addListener(() => {
   chrome.storage.sync.get('clips', item => {
     const memoryLength = Object.keys(item).length;
     if (memoryLength > 0 && item['clips']) {
-      chrome.browserAction.setBadgeText({
-        text: `${Object.keys(item['clips']).length}`,
-      });
+      showBadge(`${Object.keys(item['clips']).length}`);
     } else {
-      chrome.browserAction.setBadgeText({ text: '' });
+      showBadge('');
     }
   });
   chrome.storage.sync.get('settings', item => {
@@ -35,64 +34,8 @@ chrome.contextMenus.create({
   contexts: ['page', 'browser_action'],
 });
 
-const checkClipExist = (oldClips, newClip) => {
-  const clips = Object.keys(oldClips);
-  return clips.every(key => oldClips[key] !== newClip);
-};
-
-class MenuClickManager {
-  static selection(info) {
-    const { selectionText } = info;
-    chrome.storage.sync.get('clips', items => {
-      const allClips = items['clips'];
-      const clipExist = typeof allClips === 'object';
-      const itemToClip = selectionText.trim();
-      const itemToClipIsValid = selectionText.trim() !== '';
-      if (
-        clipExist
-        && itemToClipIsValid
-        && checkClipExist(allClips, selectionText)
-      ) {
-        const maxId = Math.max(...Object.keys(allClips));
-        const clips = {
-          ...items['clips'],
-          [maxId + 1]: itemToClip,
-        };
-        chrome.storage.sync.set({ clips });
-      } else if (!clipExist && itemToClipIsValid) {
-        chrome.storage.sync.set({ clips: { 1: selectionText } });
-      }
-    });
-  }
-
-  static page() {
-    const URL = 'https://github.com/Oluwasegun-AA/MultiClip';
-    chrome.tabs.create({ url: URL });
-  }
-
-  static lightTheme() {
-    return chrome.storage.sync.get('settings', items => {
-      const settings = { ...items['settings'], theme: 'light' };
-      chrome.storage.sync.set({ settings });
-    });
-  }
-
-  static darkTheme() {
-    return chrome.storage.sync.get('settings', items => {
-      const settings = { ...items['settings'], theme: 'dark' };
-      chrome.storage.sync.set({ settings });
-    });
-  }
-
-  static deleteItems() {}
-
-  static clearClippedItems() {
-    return chrome.storage.sync.remove('clips');
-  }
-}
-
 // attach listeners
 chrome.contextMenus.onClicked.addListener((info, tab) => {
   const { menuItemId } = info;
-  MenuClickManager[menuItemId](info, tab);
+  ContextMenuManager[menuItemId](info, tab);
 });
