--- src/client/scripts/options.js
@@ -1,1 +1,0 @@
-
